Reorganise code into class with methods
All of the awful setup stuff is now done in the constructor (is there a
better place?) and the main function delegates to 3 new functions:

- events() <- checks user input events
- loop()   <- handles game logic
- render() <- draws stuff to the screen

To be able to use things throughout these I had to basically write self.
in front of every variable, which is annoying but they should probably
also be turned into classes themselves anyway.

OOP FTW!!1