Consider image width when checking screen right bounds
So far you could appear to move off the screen because the check
compares the location of the animal to the edge of the screen but the
point used for the animal is on the left-hand side of the image, so it
keeps moving to the right until the last pixel is on the screen.

Subtracting the image width from the screen width fixes this.