Refactor animal as sub-class of Sprite
This will be needed for collision detection but one convenient thing is
already that a bunch of re-centring logic and manual movement maths can
now be replaced by much simpler Rects.

Rect's centre updates when you move it, it has nice semantic field names
like x & y instead of [0] & [1], as well as width and height and it has
move methods which you can use instead of counting coordinates yourself.