Only deduct points on food drop if no animal would've eaten it
Store which animal the food is for as an array of animals in a property
of the food.  This makes checking if it's completely inedible as easy
as checking the array length, if it's zero-length then it's for neither
animal (so far that's broccoli) and we don't count it as negative points
if it hits the floor.