Lose double-points if you eat the wrong food
By checking if the animal that collided with the food is in the list of
"for what" animals the food is, we can see whether that animal should
have eaten it or not.  You score a point if the animal is right but lose
double-points if the animal that ate it is the wrong one.