Freeze the head for a few moments if it eats the wrong food
This is done with another ticker.  Its counter doesn't have an event
though, so I moved it to the Sprite class's update() function and call
that every tick of the game.

If an animal eats the wrong food, calling the freeze() function sets
state to frozen and bumps up the ticker which will be decreased to zero
in the update() function.

We might want to play around with the FREEZE_TICK_TIME constant to get
it to a value that makes the game challenging/fun.